Add loyalty program features banner and refactor coupon handling for improved user experience

+++ FruitShop Razor Pages/Controllers/OrderController.cs
@@ -8,6 +8,6 @@ namespace FruitShop_Razor_Pages.Controllers;
 public class OrderController : ControllerBase
 {
     [HttpGet("calculate-loyalty-points/{totalAmount:long}")]
-    public int CalculateLoyaltyPoints(long totalAmount) =>
+    public long CalculateLoyaltyPoints(long totalAmount) =>
         BusinessRuleConstants.LoyaltyPoint.CalculateLoyaltyPoints(totalAmount);
 }
