+++ Repository/Models/Orders/Order.cs
@@ -21,7 +21,7 @@ public class Order
 
     [Required] public required long TotalAmount { get; set; }
 
-    [Required] public required int LoyaltyPointsEarned { get; set; }
+    [Required] public required long LoyaltyPointsEarned { get; set; }
 
     [Required] public required ShippingAddressSnapshot ShippingAddressSnapshot { get; set; }
